core/migration: Add IsZero method to MinionReports
To support detection of empty values.
